requirements.txt
# ============================================================================
# Requirements for Neurodevelopmental Outcome Predictor
# Python 3.8+
# ============================================================================

# Core scientific computing
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# Deep learning
torch>=1.10.0
torchvision>=0.11.0

# Medical imaging
nibabel>=3.2.1
nilearn>=0.9.0
scikit-image>=0.19.0

# Machine learning utilities
scikit-learn>=1.0.0
imbalanced-learn>=0.9.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.3.0

# Progress bars and utilities
tqdm>=4.62.0
pyyaml>=5.4.0
colorama>=0.4.4

# Network analysis
networkx>=2.6.0

# Statistical analysis
statsmodels>=0.13.0
pingouin>=0.5.0

# Image processing
opencv-python>=4.5.0
Pillow>=8.3.0

# Utilities
python-dateutil>=2.8.0
pytz>=2021.3

# Optional: For PDF report generation
# reportlab>=3.6.0
# weasyprint>=54.0

# Optional: For TensorBoard logging
# tensorboard>=2.7.0
# tensorboardX>=2.4.0

# Development tools (optional)
# jupyter>=1.0.0
# ipython>=7.28.0
# black>=21.9b0
# flake8>=4.0.0
# pytest>=6.2.0
